% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellPolys.R
\name{cellPolys}
\alias{cellPolys}
\title{Get polygons for specified cell IDs}
\usage{
cellPolys(
  polys,
  cell_ids = NULL,
  transcript_df,
  overwrite = FALSE,
  type = "chull",
  alpha = 0.1
)
}
\arguments{
\item{polys}{Polygons list to be added to}

\item{cell_ids}{Vector of cell IDs. If NULL, will calc polys for all cells in transcript_df}

\item{transcript_df}{Data frame of transcript info. Needs these columns: "cell_id", "x", "y".}

\item{overwrite}{Logical, for whether to overwrite polygons that have already been calculated}

\item{type}{Either "ahull" or "chull". The latter is faster but limited to convex shapes.}

\item{alpha}{Passed to alphahull::ahull(). Smaller makes for more nuanced polygons.}
}
\value{
A list of polygons with the entries for the cell IDs filled in
}
\description{
Get polygons for specified cell IDs
}
