% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmx_nephritis}
\alias{cosmx_nephritis}
\title{Small example CosMx data from a lupus nephritis sample}
\format{
A list with the following elements:
\itemize{
\item transcript_df A data frame holding transcript data
\item celltype A vector of cell types
\item cellcols A vector of colors for the cell types
}
}
\usage{
cosmx_nephritis
}
\description{
A 0.25-FOV excerpt from a 1000-plex CosMx run on a lupus nephritis biopsy.
}
\keyword{datasets}
