% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawPolys.R
\name{drawPolys}
\alias{drawPolys}
\title{Draw polygons}
\usage{
drawPolys(
  polys,
  cell_ids = NULL,
  col = "#00008B80",
  border = NA,
  transcript_df = NULL,
  type = "chull",
  alpha = 0.1,
  add = FALSE,
  outputpolys = FALSE,
  ...
)
}
\arguments{
\item{polys}{List of polygons}

\item{cell_ids}{Vector of cell IDs}

\item{col}{Vector of colors, aligned to the cell_IDs}

\item{border}{Vector of border colors, aligned to the cell_IDs}

\item{transcript_df}{Data frame of transcript info. Needs these columns: "cell_id", "x", "y".}

\item{type}{Either "ahull" or "chull". The latter is faster but limited to convex shapes.}

\item{alpha}{Passed to alphahull::ahull(). Smaller makes for more nuanced polygons.}

\item{add}{Logical, for whether to open a new plot. If FALSE then polys are drawn atop existing plot.}

\item{outputpolys}{Whether to return the polygons list (useful if you've calculated more polygons within the function call).}

\item{...}{Arguments passed to plot(), in the event that add == FALSE}
}
\value{
A list of polygons with the entries for the cell IDs filled in
}
\description{
Draws polygons for given cell IDs, calculating them from the transcript data frame if needed
}
