% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellPolys.R
\name{initPolys}
\alias{initPolys}
\title{Initialize an empty polygons list}
\usage{
initPolys(cell_ids = NULL, transcript_df = NULL)
}
\arguments{
\item{cell_ids}{Vector of cell IDs}

\item{transcript_df}{Data frame of transcript info. Uses the "cell_id" column to derive cell IDs.
Must provide either this or "cell_ids" argument directly.}
}
\description{
Initialize an empty polygons list
}
