% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notredamecols}
\alias{notredamecols}
\title{Colors for stained glass plots}
\format{
A vector of colors
}
\usage{
notredamecols
}
\description{
Colors matching the north rose window of NotreDame
}
\keyword{datasets}
