% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StainedGlass.R
\name{stainedGlassPlot}
\alias{stainedGlassPlot}
\title{Make a stained glass plot}
\usage{
stainedGlassPlot(
  polys,
  transcript_df,
  celltype,
  transcript_cex = 0.3,
  cellcols = notredamecols
)
}
\arguments{
\item{polys}{List of polygons}

\item{transcript_df}{Data frame of transcript info. Needs these columns: "cell_id", "x", "y".}

\item{celltype}{Vector of cell types, with names aligned to the names of polys}

\item{transcript_cex}{Point size of transcripts. Default works well for a 1/4 FOV view}

\item{cellcols}{Vector of colors, named by celltype}
}
\description{
Uses drawPolys to make a plot resembling stained glass
}
